Switch to webpack instead of brunch.
Brunch seems to be having extreme problems on travis, so I'm trying out
webpack.  Webpack is also really slow to compile on travis, but not
quite as slow as brunch, which times out a good percentage of the time.

This updates most of the config to using webpack, following this guide:
http://matthewlehner.net/using-webpack-with-phoenix-and-elixir/
Also with help from
https://www.terakilobyte.com/using-webpack-with-phoenix-and-elm/
and
http://www.elm-tutorial.org/en/04-starting/03-webpack.html
for some of the elm stuff.

Also tried out node 6 on travis, just incase it helped with speed.  It
doesn't seem to have done that.